06-mcp-http/requirements.txt
# Core dependencies
fastmcp>=0.1.0
langgraph>=0.2.0
langchain>=0.3.0
langchain-anthropic>=0.3.0
anthropic>=0.39.0

# Serialization support
pyyaml>=6.0.1

# Async HTTP client
httpx>=0.27.0

# Environment management
python-dotenv>=1.0.0

# Type hints and validation
pydantic>=2.0.0